Add modification time change check of the file

+++ importtosql.py
@@ -1,9 +1,12 @@
+import os.path
+import time
 import pandas as pd
 from sqlalchemy import create_engine
 
+
 with open('config.txt') as f:
-    contents = f.read()
-    print(contents)
+    path = f.read()
+    print(path)
 
 #Sec
 svr_name = 'LDWSWISQLVAL01'
@@ -11,14 +14,26 @@ db_name = 'MaintenanceReport'
 u_name = 'PowerBI'
 u_pass = '111098tdg'
 
-df = pd.read_excel(contents)
 
-df['Początek zakłócenia'] = df['Początek zakłócenia'].astype('string')
-df['Pocz. zakłóc. (godz.)'] = df['Pocz. zakłóc. (godz.)'].astype('string')
-df['Koniec zakłóc.(godz.)'] = df['Koniec zakłóc.(godz.)'].astype('string')
+last_time = time.ctime(os.path.getmtime(path))
+print(last_time)
 
-engine = create_engine("mssql+pyodbc://"+u_name+":"+u_pass+"@"+svr_name+"/MaintenanceReport?driver=SQL+Server", fast_executemany=True)
-engine.execute("delete from MaintenanceData")
+while(True):
+    if time.ctime(os.path.getmtime(path)) != last_time:
+        df = pd.read_excel(path)
 
-df.to_sql('MaintenanceData', if_exists='append', con=engine, index=False)
+        df['Początek zakłócenia'] = df['Początek zakłócenia'].astype('string')
+        df['Pocz. zakłóc. (godz.)'] = df['Pocz. zakłóc. (godz.)'].astype('string')
+        df['Koniec zakłóc.(godz.)'] = df['Koniec zakłóc.(godz.)'].astype('string')
 
+        engine = create_engine(
+            "mssql+pyodbc://" + u_name + ":" + u_pass + "@" + svr_name + "/MaintenanceReport?driver=SQL+Server",
+            fast_executemany=True)
+        engine.execute("delete from MaintenanceData")
+
+        df.to_sql('MaintenanceData', if_exists='append', con=engine, index=False)
+
+        print('New file send to SQL!')
+
+        break
+    time.sleep(10)
